Added config comments and variable expansion
If we want Kerberos.NET to be completely compliant with MIT Kerberos
defaults, then we can change some of the paths. The location of
krb5.conf is usually system-wide (/etc/krb5.conf) and the kdc.conf is in
the $LOCALSTATEDIR/krb5kdc. /etc isn't a user configuration location so
I changed it to $XDG_CONFIG_HOME, and I added the ability to expand
variables set in the krb5.conf. That can allow putting the krb5ccache in
the home directory with %HOME%. We should consider honoring the
%KRB5CCNAME% environment variable.